git-svn-id: https://svn.code.sf.net/p/quake2xp/code@1189 d83f5309-02f3-4a42-a5f4-12e6af65da40

+++ trunk/client/ref.h
@@ -278,7 +278,7 @@ typedef enum {
 	it_bump,
 } imagetype_t;
 
-typedef unsigned long int uint64;
+typedef unsigned long long uint64;
 
 typedef struct image_s {
